Combined AllBuildingsCard and SavedHomesCard into one BuildingsCard component
